+++ internal/genre/repository.go
@@ -9,7 +9,7 @@ import (
 type Repo interface {
 	Create(ctx context.Context, genre *models.Genre) (*models.Genre, error)
 	GetAll(ctx context.Context) ([]*models.Genre, error)
-	GetAllByArtistID(ctx context.Context, artistID int) ([]*models.Genre, error)
-	GetAllByAlbumID(ctx context.Context, albumID int) ([]*models.Genre, error)
-	GetAllByTrackID(ctx context.Context, trackID int) ([]*models.Genre, error)
+	GetAllByArtistID(ctx context.Context, artistID uint64) ([]*models.Genre, error)
+	GetAllByAlbumID(ctx context.Context, albumID uint64) ([]*models.Genre, error)
+	GetAllByTrackID(ctx context.Context, trackID uint64) ([]*models.Genre, error)
 }
